refactor: create AppBarWithDrawer
