+++ src/pages/detail/ProductDetail.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import AppBarWithDrawer from "../../components/shared/AppBarWithDrawer";
+
+const ProductDetail = () => {
+	return (
+		<>
+			<AppBarWithDrawer/>
+		</>
+	)
+};
+
+export default ProductDetail;
